pyproject.toml
[tool.poetry]
name = "tusdatos"
version = "0.1.0"
description = "Technical test for the position of backend developer for tusdatos.co"
authors = ["Gabriel Morales <dev@darklock.maskmy.id>"]
readme = "README.md"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.11"
httpx= "0.27.0"
motor = "3.3.2"
pydantic = "2.6.3"
pydantic-settings = "2.2.1"
fastapi = "0.110.0"
uvicorn = {extras = ["standard"], version = "0.28.0"}
python-jose = {extras = ["cryptography"], version = "3.3.0"}
passlib = {extras = ["bcrypt"], version = "1.7.4"}
sqlalchemy = {extras = ["asyncio"], version = "2.0.28"}
asyncpg = "0.29.0"
alembic = "1.13.1"
python-multipart = "0.0.9"

[tool.poetry.group.dev.dependencies]
black = "24.1.1"
ipython = "8.22.2"
isort = "5.13.2"
jedi-language-server = "0.41.2"
ruff-lsp = "0.0.52"

[tool.isort]
profile = "black"

[tool.ruff.isort]
combine-as-imports = true

[tool.black]
line-length = 100
